+++ Changelog.md
@@ -4,6 +4,7 @@
 * Created common comms management logic in HeLMS
 * Steersman comms control option added
 * Added separate asset pool activity detection for non-air groups (CPS and ROC now compatible with these groups)
+* Added filters to CPS controllable from the comms menu. TODO: document it!
 ## 1.7a
 * Mission examples added
 * Added late activation handling for unit_repairman